open menu directly next to its parent.
Originally it opened directly on top of its parent, then
b160e08db9a866d4d1b46a12d0f5fde854c7b4ac introduced a bug
that caused it to open at (0,0). Now we open it to
the right of the parent because that's where we actually
want it to be in all flux maker tools uses.